Fix race condition in synchronous requests
Whenever invoked as part of the `magit-post-refresh-hook', the request
function from request.el was returning a `request-response' object
whose status-code and data slots were not yet populated, even with the
`sync' parameter set.

I believe this results from a race condition in which the process
sentinel for the curl command doesn't get a chance to run before the
process status is updated, causing the `request' function to return
before its callbacks can update the state of the `request-response'
object.

Rather than rely on the `sync' option (which the request.el docs warn
against, anyway), I've simply implemented a synchronous wrapper which
invokes `sit-for' (with a timeout) so that the process sentinel for
the curl command is given an opportunity to run and populate the
response.