Add review-centric view
Replace the initial, diff-centric view with a review-centric view. This
view is modeled after the Github web interface, in which reviews are
shown sequentially and each comment thread in the review is displayed
with a few lines of the diff for context.

Unlike the Github interface, this interfaces allows the user to jump
from a comment directly to the corresponding line in the diff (and from
a line in the diff to the corresponding line in the file).