Remove carriage-returns from Github responses
Github seems to use Windows-style new-lines, i.e., `\r\n`. The
carriage-returns are displayed as `^M` in Emacs, which is ugly.

This PR recursively strips any carriage-returns from the response from
Github, unless the response is not a list, which we're using as a
proxy for "not a diff." We should probably check against the "Accept"
type in the request headers instead. I started to go down that route
but then got lazy.