Refactoring: represent review as struct instead of alist
Return a list of structs from `magit-gh-list-reviews' instead of a
list of alists. We're now consistent in our internal representation of
reviews as using the `magit-gh-review' struct.

I also switched to using hashtables in the implementation of
`magit-gh-list-reviews'. Dealing with three levels of nesting in the
grouped list of alists was getting hard to reason about.