Add minor mode for editing comments
I'm not happy with this approach and I plan to refactor it.

Currently, magit-gh-comments mode is initialized with a PR and an
"action" keyword arg, one of `:view-pr` or `:submit-review`. We
populate the buffer differently depending on the action.

Instead, I want to split this into two separate minor modes.

Another thing I don't like about this is the way the prose minor- mode
replaces the content of a magit-section. It took a lot of fiddling to
get the `start` and `end` of the magit-section populated correctly,
and I'm not convinced the current implementation is very robust. I
think magit sections just aren't meant to be mutable.

A cleaner approach would be to just to initialize the buffer with a
callback that runs when editing is complete. That hook will update the
pending review draft and then refresh the src buffer, which will
automatically pick up and redisplay the new content.