Allow replies to comments. Display comment threads
Tests are failing - I'm going to follow up with a commit to fix
them (I refactored the test infrastructure for making assertions about
the magit-section tree, but don't want to make this commit too big).

Note that there is one known bug in which pending comments aren't
displayed when submitting a review. I'll follow up with a unit test to
expose the bug and a fix in a subsequent commit.