Post replies immediately. Represent comments as threads (#46)
Model comment threads as a top-level field on a review. The function magit-gh-list-comments is
removed, since we almost never work with individual comments. We'll use magit-gh-list-reviews and access the threads field of a review any time we need access to comments.

I've removed the ability to save reply-to comments in a draft. Not only was it tricky to implement, but the UX was kind of confusing. This is how the Github web UI handles replies, anyway.